chore: monorepo restructure — update configs, stories imports, and package names
