+++ packages/core/README.md
@@ -1,0 +1,16 @@
+# @falkordb/ui
+
+Shared React UI component library for FalkorDB applications.
+
+## Install
+
+```bash
+npm install @falkordb/ui
+```
+
+## Usage
+
+```tsx
+import { Button, Card } from '@falkordb/ui'
+import '@falkordb/ui/theme/tokens.css'
+```
