+++ packages/core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@falkordb/ui-core",
+  "name": "@falkordb/ui",
   "version": "0.1.0",
   "description": "Shared React UI component library for FalkorDB applications",
   "type": "module",
