+++ stories/Button.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { Button } from '../src/components/ui/button'
+import { Button } from '../packages/core/src/components/ui/button'
 import { Mail } from 'lucide-react'
 
 const meta: Meta<typeof Button> = {
